chore: removed unnecessary conversion from long to int and back again

+++ src/main/kotlin/nl/info/zac/app/inboxdocument/InboxDocumentRestService.kt
@@ -60,7 +60,7 @@ class InboxDocumentRestService @Inject constructor(
         val restInboxDocuments = inboxDocuments.toRestInboxDocuments(informatieobjectTypeUUIDs)
         return RESTResultaat<RestInboxDocument>(
             restInboxDocuments,
-            inboxDocumentService.count(listParameters).toLong()
+            inboxDocumentService.count(listParameters)
         )
     }
